+++ packages/insomnia/src/ui/components/dropdowns/git-sync-dropdown.tsx
@@ -29,7 +29,6 @@ import { GitLogModal } from '../modals/git-log-modal';
 import { GitStagingModal } from '../modals/git-staging-modal';
 
 type Props = ReturnType<typeof mapDispatchToProps> & {
-  handleGitBranchChanged: (branch: string) => void;
   workspace: Workspace;
   vcs: GitVCS;
   gitRepository: GitRepository | null;
@@ -64,7 +63,7 @@ class GitSyncDropdown extends PureComponent<Props, State> {
   }
 
   async _refreshState(otherState?: Record<string, any>) {
-    const { vcs, workspace, handleGitBranchChanged } = this.props;
+    const { vcs, workspace } = this.props;
     const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspace._id);
 
     // Clear cached items and return if no state
@@ -101,7 +100,6 @@ class GitSyncDropdown extends PureComponent<Props, State> {
       cachedGitLastAuthor,
       cachedGitLastCommitTime,
     });
-    handleGitBranchChanged(branch);
   }
 
   async _handleOpen() {
